core/local/chokidar: Extract ./prepare_events from ./watcher
